Migrate avatar from /docs/components/avatar to /components/avatar
Port all examples (Basic, With Fallback, Group) from the old doc page
to the new reference page, remove the old page and its route, and
update E2E tests to target the new URL with scoped selectors.
